# [A2-798] opa update time take two: flatten internal structure