… components/authz-service/engine/opa/example_v2/policies.json
… components/authz-service/engine/opa/example_v2/store-flat.json
… components/authz-service/engine/opa/example_v2/store-pretty.json
… components/authz-service/engine/opa/opa_v2_test.go
… components/authz-service/engine/opa/policy.bindata.go
… components/authz-service/engine/opa/policy/authz_v2_test.rego