### :nut_and_bolt: Description

Since introducing the IAM v2 policy structure, featuring multiple statements in a policy, we've never benchmarked what this structure means for our performance. Internal to OPA, we've fed in the nested data (see [this examples file](https://github.com/chef/automate/pull/362/files#diff-5a7e9b97ddb89aeded901363feb0d1aa)) and had our rego code process it. Adding a benchmark in #323, we've found that updating the partial result with this takes rather long.

Incidentally, we don't need that information to make a policy decision: only the statements matter. Also, we've given the statements unique IDs, so that they can be keys in an object (see [the perf topic in OPA's FAQ](https://github.com/open-policy-agent/opa/blob/master/docs/content/faq.md#high-performance-policy-decisions)). With *this change*, we're flattening the internal structure, [see this example file](https://github.com/chef/automate/pull/362/files#diff-91ff39bcdd0ac49faefdd73cba3f67b1), and process that instead.

<details><summary>I retract that benchmark. Will have to look closer</summary>


 The benchmark shows that the change is dramatic:

```
name                                                 old time/op    new time/op    delta
InitPartialResultV2/default_policies,_run_0_times-8     1.34s ±17%     0.11s ± 4%  -91.49%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_1_times-8     2.21s ± 6%     0.23s ± 8%  -89.65%  (p=0.000 n=10+9)
InitPartialResultV2/default_policies,_run_2_times-8     3.14s ± 4%     0.34s ± 3%  -89.06%  (p=0.000 n=10+9)
InitPartialResultV2/default_policies,_run_3_times-8     4.30s ± 1%     0.51s ±17%  -88.19%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_4_times-8     5.34s ± 1%     0.65s ±20%  -87.85%  (p=0.000 n=10+10)
InitPartialResultV2/0_dummy_policies-8                  305ms ± 1%     310ms ± 3%   +1.67%  (p=0.011 n=9+9)
InitPartialResultV2/25_dummy_policies-8                 330ms ± 2%     338ms ± 3%   +2.36%  (p=0.010 n=10+9)
InitPartialResultV2/50_dummy_policies-8                 360ms ± 2%     358ms ± 8%     ~     (p=0.143 n=10+10)
InitPartialResultV2/75_dummy_policies-8                 395ms ± 1%     361ms ± 2%   -8.53%  (p=0.000 n=10+10)
InitPartialResultV2/100_dummy_policies-8                439ms ± 4%     395ms ± 2%  -10.07%  (p=0.000 n=10+10)

name                                                 old alloc/op   new alloc/op   delta
InitPartialResultV2/default_policies,_run_0_times-8     259MB ± 0%      21MB ± 0%  -91.96%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_1_times-8     468MB ± 0%      42MB ± 0%  -91.11%  (p=0.000 n=10+8)
InitPartialResultV2/default_policies,_run_2_times-8     677MB ± 0%      64MB ± 0%  -90.56%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_3_times-8     887MB ± 0%      84MB ± 5%  -90.53%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_4_times-8    1.10GB ± 0%    0.10GB ± 0%  -90.51%  (p=0.000 n=10+9)
InitPartialResultV2/0_dummy_policies-8                 65.9MB ± 0%    63.0MB ± 0%   -4.36%  (p=0.000 n=10+8)
InitPartialResultV2/25_dummy_policies-8                70.3MB ± 0%    67.1MB ± 0%   -4.61%  (p=0.000 n=10+9)
InitPartialResultV2/50_dummy_policies-8                75.8MB ± 0%    71.6MB ± 0%   -5.51%  (p=0.000 n=10+10)
InitPartialResultV2/75_dummy_policies-8                82.2MB ± 0%    76.6MB ± 0%   -6.89%  (p=0.000 n=10+10)
InitPartialResultV2/100_dummy_policies-8               89.8MB ± 0%    82.1MB ± 0%   -8.60%  (p=0.000 n=9+10)

name                                                 old allocs/op  new allocs/op  delta
InitPartialResultV2/default_policies,_run_0_times-8     8.68M ± 0%     0.85M ± 0%  -90.21%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_1_times-8     15.6M ± 0%      1.7M ± 0%  -89.12%  (p=0.000 n=7+9)
InitPartialResultV2/default_policies,_run_2_times-8     22.6M ± 0%      2.6M ± 0%  -88.45%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_3_times-8     29.5M ± 0%      3.4M ± 4%  -88.38%  (p=0.000 n=10+10)
InitPartialResultV2/default_policies,_run_4_times-8     36.4M ± 0%      4.2M ± 0%  -88.33%  (p=0.000 n=10+10)
InitPartialResultV2/0_dummy_policies-8                  2.46M ± 0%     2.36M ± 0%   -4.11%  (p=0.000 n=10+8)
InitPartialResultV2/25_dummy_policies-8                 2.60M ± 0%     2.49M ± 0%   -4.21%  (p=0.000 n=10+10)
InitPartialResultV2/50_dummy_policies-8                 2.78M ± 0%     2.65M ± 0%   -4.76%  (p=0.000 n=10+10)
InitPartialResultV2/75_dummy_policies-8                 2.99M ± 0%     2.82M ± 0%   -5.64%  (p=0.000 n=10+10)
InitPartialResultV2/100_dummy_policies-8                3.23M ± 0%     3.01M ± 0%   -6.78%  (p=0.000 n=10+10)
```
and our passing test show that the new way works just the same as the old way did.

</details>

ℹ️ Should we ever need the policy ID, for example to return it to indicate *why something was denied*, it could also be passed on to the statements, [see this comment](https://github.com/chef/automate/pull/362/files#diff-2d317bb33ff7babca7e1290b0fb1dcf2R663).

### :+1: Definition of Done

- Everything works like before, but updating policies/roles/rules does not take very long anymore.

### :athletic_shoe: Demo Script / Repro Steps

- deployinate, `rebuild components/authz-service`

### :chains: Related Resources

- [A2-798](https://chefio.atlassian.net/browse/A2-798)

### :white_check_mark: Checklist

- [X] Necessary tests added/updated?
- [ ] Necessary docs added/updated?
- [X] Code actually executed?
- [X] Vetting performed (unit tests, lint, etc.)?